refactor: useBeerQuery naming changed
