+++ src/components/home/HomeTemplate.tsx
@@ -1,6 +1,5 @@
 import {Box, Container} from "@chakra-ui/react";
 import {useEffect} from "react";
-import {useTopBeersQuery} from "../../../hooks/useTopBeersQuery";
 import {SignUpType} from "../../../interface/types";
 import {BeerResponseType} from "../../../typedef/server/beer";
 import {BlankHeader} from "../shared/Headers/BlankHeader";
@@ -8,6 +7,7 @@ import {CommonBeersList} from "./CommonBeersList/CommonBeersList";
 import {LoggedInBeersList} from "./LoggedInBeersList/LoggedInBeersList";
 import SearchInputHome from "./SearchInputHome";
 import {WelcomeTextContent} from "./WelcomeText";
+import {useTopBeersQuery} from "@/../hooks/useBeerQuery";
 
 interface HomeTemplateProps {
   userInfo?: SignUpType;
